Formatted code and implemented missing functionality. There now are helper functions for setting different data types on vertex and index buffers and one for setting 2D texture data. There is no simple way to set only parts of a texture and no simple way to set cube texture data. Right now this has to be done using the provided pitch and unsafe.Pointer in LOCKED_RECT and LOCKED_BOX. Helper functions will be created in the future as needed.

+++ samples/simple_glow/main.go
@@ -46,7 +46,7 @@ func main() {
 		check(device.Clear(
 			nil,
 			d3d9.CLEAR_TARGET,
-			d3d9.ColorRGB(uint(red), 92, 128),
+			d3d9.ColorRGB(uint8(red), 92, 128),
 			0,
 			0,
 		))
